requirements.txt
flask
flask-sqlalchemy
flask-cors
pandas
openpyxl
pdfplumber
python-dotenv
flask-jwt-extended
psycopg2-binary
werkzeug
fpdf2
